requirements.txt
numpy==1.26.2
pandas==2.1.3
pytest==7.4.3
PyYAML==6.0.1
rdt==1.9.0
Requests==2.31.0
scikit_learn==1.3.2
setuptools==68.0.0
